+++ src/ais/mod.rs
@@ -25,6 +25,8 @@ pub mod vdm_t5;
 #[doc(hidden)]
 pub mod vdm_t6;
 #[doc(hidden)]
+pub mod vdm_t9;
+#[doc(hidden)]
 pub mod vdm_t18;
 #[doc(hidden)]
 pub mod vdm_t19;
@@ -38,6 +40,7 @@ pub mod vdm_t27;
 use super::*;
 pub use vdm_t4::{BaseStationReport};
 pub use vdm_t6::{BinaryAddressedMessage};
+pub use vdm_t9::{StandardSarAircraftPositionReport};
 pub use vdm_t21::{AidToNavigationReport, NavAidType};
 
 // -------------------------------------------------------------------------------------------------